Much cleanups
